[REFACTOR] total score 반환 형태 수정
- Integer에서 Long으로 수정
- 기본 JPQL에서 제공하는 SUM, MIN, MAX 등 반환 형태가 Long타입임

+++ src/main/java/org/example/domain/rank/controller/dto/RankDataResponse.java
@@ -8,5 +8,5 @@ import org.example.domain.member.domain.Member;
 @AllArgsConstructor
 public class RankDataResponse {
     private Member member;
-    private Integer totalScore;
+    private Long totalScore;
 }
